Address requested changes. - Remove Translation component and unused code. - Use Promise.all for update requests. - Add class method for wrkflow & proj GET req's. - Move WorkflowDefaultDialog to admin folder.
